# Conflicts:
#	DEPS
#	webrtc/api/peerconnection.cc
#	webrtc/api/peerconnectionfactory.cc
#	webrtc/api/peerconnectionfactory.h
#	webrtc/api/rtpsenderreceiver_unittest.cc
#	webrtc/api/videocapturertracksource.cc
#	webrtc/api/videosource.h
#	webrtc/api/videosourceinterface.h
#	webrtc/api/videosourceproxy.h
#	webrtc/api/videotrack.cc
#	webrtc/api/videotrackrenderers.cc
#	webrtc/api/videotrackrenderers.h
#	webrtc/base/base.gyp
#	webrtc/base/base_tests.gyp
#	webrtc/base/bytebuffer_unittest.cc
#	webrtc/base/messagequeue_unittest.cc
#	webrtc/base/nethelpers.cc
#	webrtc/base/platform_thread.cc
#	webrtc/base/scoped_ptr.h
#	webrtc/base/timeutils.cc
#	webrtc/base/timeutils.h
#	webrtc/build/common.gypi
#	webrtc/build/gyp_webrtc
#	webrtc/common_audio/signal_processing/include/spl_inl.h
#	webrtc/common_types.h
#	webrtc/common_video/incoming_video_stream.cc
#	webrtc/media/base/executablehelpers.h
#	webrtc/media/base/mediaengine.h
#	webrtc/media/base/testutils.cc
#	webrtc/media/base/videocapturer.h
#	webrtc/media/devices/devicemanager.cc
#	webrtc/media/engine/webrtcvideocapturer.cc
#	webrtc/media/engine/webrtcvoiceengine.cc
#	webrtc/media/media.gyp
#	webrtc/media/media_tests.gypi
#	webrtc/modules/audio_coding/neteq/decoder_database.cc
#	webrtc/modules/audio_coding/neteq/neteq_impl.cc
#	webrtc/modules/audio_coding/neteq/normal.cc
#	webrtc/modules/audio_device/audio_device.gypi
#	webrtc/modules/audio_device/audio_device_impl.cc
#	webrtc/modules/audio_device/include/audio_device.h
#	webrtc/modules/audio_processing/aecm/aecm_core_neon.cc
#	webrtc/modules/audio_processing/audio_processing_impl.cc
#	webrtc/modules/modules.gyp
#	webrtc/modules/remote_bitrate_estimator/overuse_detector.cc
#	webrtc/modules/rtp_rtcp/source/h264_bitstream_parser.cc
#	webrtc/modules/video_capture/test/video_capture_unittest.cc
#	webrtc/modules/video_capture/windows/video_capture_factory_windows.cc
#	webrtc/modules/video_coding/codecs/vp8/screenshare_layers_unittest.cc
#	webrtc/modules/video_coding/codecs/vp8/vp8_impl.h
#	webrtc/modules/video_coding/generic_encoder.cc
#	webrtc/modules/video_coding/generic_encoder.h
#	webrtc/modules/video_coding/timing.h
#	webrtc/modules/video_coding/utility/quality_scaler.cc
#	webrtc/modules/video_coding/utility/quality_scaler.h
#	webrtc/modules/video_coding/utility/quality_scaler_unittest.cc
#	webrtc/modules/video_coding/video_sender.cc
#	webrtc/modules/video_render/video_render.gypi
#	webrtc/modules/video_render/windows/video_render_windows_impl.cc
#	webrtc/modules/video_render/windows/video_render_windows_impl.h
#	webrtc/p2p/p2p_tests.gypi
#	webrtc/pc/channelmanager.cc
#	webrtc/pc/channelmanager.h
#	webrtc/pc/pc_tests.gypi
#	webrtc/system_wrappers/source/condition_variable_unittest.cc
#	webrtc/system_wrappers/source/metrics_unittest.cc
#	webrtc/system_wrappers/source/tick_util.cc
#	webrtc/test/histogram.cc
#	webrtc/test/test.gyp
#	webrtc/test/testsupport/fileutils.cc
#	webrtc/video/payload_router_unittest.cc
#	webrtc/video/vie_channel.cc
#	webrtc/video/vie_channel.h
#	webrtc/video_encoder.h
#	webrtc/voice_engine/channel.cc
#	webrtc/voice_engine/channel.h
#	webrtc/voice_engine/voe_base_impl.cc
#	webrtc/webrtc_tests.gypi

+++ webrtc/api/androidtests/src/org/webrtc/NetworkMonitorTest.java
@@ -24,7 +24,6 @@ import android.net.wifi.WifiManager;
 import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
-import android.telephony.TelephonyManager;
 import android.test.ActivityTestCase;
 import android.test.UiThreadTest;
 import android.test.suitebuilder.annotation.MediumTest;
@@ -72,7 +71,7 @@ public class NetworkMonitorTest extends ActivityTestCase {
     // Dummy implementations to avoid NullPointerExceptions in default implementations:
 
     @Override
-    public int getDefaultNetId() {
+    public long getDefaultNetId() {
       return INVALID_NET_ID;
     }
 
@@ -127,7 +126,7 @@ public class NetworkMonitorTest extends ActivityTestCase {
     public void onNetworkConnect(NetworkInformation networkInfo) {}
 
     @Override
-    public void onNetworkDisconnect(int networkHandle) {}
+    public void onNetworkDisconnect(long networkHandle) {}
   }
 
   private static final Object lock = new Object();
